Corrected path for location of bootstrap source to fix the "Missing version.properties" error at run-time
