+++ core/bootstrap/src/main/config/META-INF/jboss-beans.xml
@@ -198,8 +198,8 @@
     <!-- ==================================================================== -->
     <!-- RestComm SS7 Service -->
     <!-- ==================================================================== -->
-    <bean name="MAPSS7Service" class="org.mobicents.applications.gmlc.bootstrap.SS7Service">
-        <annotation>@org.jboss.aop.microcontainer.aspects.jmx.JMX(name="org.mobicents.ss7:service=SS7Service",exposedInterface=org.mobicents.applications.gmlc.bootstrap.SS7ServiceMBean.class,registerDirectly=true)
+    <bean name="MAPSS7Service" class="org.mobicents.gmlc.server.bootstrap.SS7Service">
+        <annotation>@org.jboss.aop.microcontainer.aspects.jmx.JMX(name="org.mobicents.ss7:service=SS7Service",exposedInterface=org.mobicents.gmlc.server.bootstrap.SS7ServiceMBean.class,registerDirectly=true)
         </annotation>
         <property name="jndiName">java:/restcomm/ss7/map</property>
         <property name="stack">
